Swap bytes in P-ROM after objcopy
Currently the examples swap bytes of the P-ROM with objcopy, which
operates on each section from the ELF, rather than on the output file.
This only works when a section's size is even, which is not enforced
by the linkscript.
Use dd to swap bytes instead, to keep linkscript simple.

Closes dciabrin/ngdevkit#17