Make examples create sprites from tile #256
The examples configure the original BIOS's Eye Catcher to use
tile sprites from offset #0. 60 tiles are used for the Neo Geo logo
but tile #255 is used as well (as an empty tile).

Update all examples to create their sprite starting from tile 256
to avoid artifacts when running on real hardware.