Move png assets to gif to avoid loading issues in PyGame
When PyGame is built with support for stb_image, the latter
automatically translates paletted png to full color surfaces,
thus losing palette and pixel data.

Move all the png assets to gif, to ensure that PyGame will
always be able to load images, no matter how it is compiled
in the environment.

Ref dciabrin/ngdevkit#96